+++ Sources/Sentry/SentryUIViewControllerPerformanceTracker.m
@@ -11,10 +11,8 @@
 #    import "SentrySpanId.h"
 #    import "SentrySwift.h"
 #    import "SentryTimeToDisplayTracker.h"
-#    import "SentryTraceOrigins.h"
 #    import "SentryTracer.h"
 #    import <SentryInAppLogic.h>
-#    import <SentrySpanOperations.h>
 #    import <UIKit/UIKit.h>
 #    import <objc/runtime.h>
 
@@ -45,7 +43,7 @@
         self.inAppLogic = [[SentryInAppLogic alloc] initWithInAppIncludes:options.inAppIncludes
                                                             inAppExcludes:options.inAppExcludes];
 
-        _enableWaitForFullDisplay = NO;
+        _alwaysWaitForFullDisplay = NO;
         _dispatchQueueWrapper = SentryDependencyContainer.sharedInstance.dispatchQueueWrapper;
     }
     return self;
@@ -129,8 +127,8 @@
         NSString *name = [SwiftDescriptor getViewControllerClassName:controller];
         spanId = [self.tracker startSpanWithName:name
                                       nameSource:kSentryTransactionNameSourceComponent
-                                       operation:SentrySpanOperationUILoad
-                                          origin:SentryTraceOriginAutoUIViewController];
+                                       operation:SentrySpanOperation.uiLoad
+                                          origin:SentryTraceOrigin.autoUIViewController];
 
         // Use the target itself to store the spanId to avoid using a global mapper.
         objc_setAssociatedObject(controller, &SENTRY_UI_PERFORMANCE_TRACKER_SPAN_ID, spanId,
@@ -163,25 +161,51 @@
         return;
     }
 
-    [self.currentTTDTracker finishSpansIfNotFinished];
-
     SentryTimeToDisplayTracker *ttdTracker =
-        [[SentryTimeToDisplayTracker alloc] initForController:controller
-                                           waitForFullDisplay:self.enableWaitForFullDisplay
-                                         dispatchQueueWrapper:_dispatchQueueWrapper];
+        [self startTimeToDisplayTrackerForScreen:[SwiftDescriptor getObjectClassName:controller]
+                              waitForFullDisplay:self.alwaysWaitForFullDisplay
+                                          tracer:(SentryTracer *)vcSpan];
 
-    if ([ttdTracker startForTracer:(SentryTracer *)vcSpan]) {
+    if (ttdTracker) {
         objc_setAssociatedObject(controller, &SENTRY_UI_PERFORMANCE_TRACKER_TTD_TRACKER, ttdTracker,
             OBJC_ASSOCIATION_ASSIGN);
-        self.currentTTDTracker = ttdTracker;
-    } else {
-        self.currentTTDTracker = nil;
     }
 }
 
 - (void)reportFullyDisplayed
 {
+    SentryTimeToDisplayTracker *tracker = self.currentTTDTracker;
+    if (tracker == nil) {
+        SENTRY_LOG_DEBUG(@"No screen transaction being tracked right now.")
+        return;
+    }
+    if (!tracker.waitForFullDisplay) {
+        SENTRY_LOG_WARN(@"Transaction is not waiting for full display report. You can enable "
+                        @"`enableTimeToFullDisplay` option, or use the waitForFullDisplay "
+                        @"property in our `SentryTracedView` view for SwiftUI.");
+        return;
+    }
     [self.currentTTDTracker reportFullyDisplayed];
+}
+
+- (SentryTimeToDisplayTracker *)startTimeToDisplayTrackerForScreen:(NSString *)screenName
+                                                waitForFullDisplay:(BOOL)waitForFullDisplay
+                                                            tracer:(SentryTracer *)tracer
+{
+    [self.currentTTDTracker finishSpansIfNotFinished];
+
+    SentryTimeToDisplayTracker *ttdTracker =
+        [[SentryTimeToDisplayTracker alloc] initWithName:screenName
+                                      waitForFullDisplay:waitForFullDisplay
+                                    dispatchQueueWrapper:_dispatchQueueWrapper];
+
+    if ([ttdTracker startForTracer:tracer] == NO) {
+        self.currentTTDTracker = nil;
+        return nil;
+    }
+
+    self.currentTTDTracker = ttdTracker;
+    return ttdTracker;
 }
 
 - (void)viewControllerViewWillAppear:(UIViewController *)controller
@@ -202,8 +226,8 @@
             SENTRY_LOG_DEBUG(@"Tracking UIViewController.viewWillAppear");
             [self.tracker measureSpanWithDescription:@"viewWillAppear"
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
-                                              origin:SentryTraceOriginAutoUIViewController
+                                           operation:SentrySpanOperation.uiLoad
+                                              origin:SentryTraceOrigin.autoUIViewController
                                              inBlock:callbackToOrigin];
         };
 
@@ -229,10 +253,10 @@
 
 /**
  * According to the apple docs, see
- * https://developer.apple.com/documentation/uikit/uiviewcontroller: Not all ‘will’ callback methods
- * are paired with only a ‘did’ callback method. You need to ensure that if you start a process in a
- * ‘will’ callback method, you end the process in both the corresponding ‘did’ and the opposite
- * ‘will’ callback method.
+ * https://developer.apple.com/documentation/uikit/uiviewcontroller: Not all ‘will’ callback
+ * methods are paired with only a ‘did’ callback method. You need to ensure that if you start a
+ * process in a ‘will’ callback method, you end the process in both the corresponding ‘did’ and
+ * the opposite ‘will’ callback method.
  *
  * As stated above @c viewWillAppear doesn't need to be followed by a @c viewDidAppear. A
  * @c viewWillAppear can also be followed by a @c viewWillDisappear. Therefore, we finish the
@@ -268,8 +292,8 @@
         void (^duringBlock)(void) = ^{
             [self.tracker measureSpanWithDescription:lifecycleMethod
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
-                                              origin:SentryTraceOriginAutoUIViewController
+                                           operation:SentrySpanOperation.uiLoad
+                                              origin:SentryTraceOrigin.autoUIViewController
                                              inBlock:callbackToOrigin];
         };
 
@@ -312,15 +336,15 @@
         void (^duringBlock)(void) = ^{
             [self.tracker measureSpanWithDescription:@"viewWillLayoutSubviews"
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
-                                              origin:SentryTraceOriginAutoUIViewController
+                                           operation:SentrySpanOperation.uiLoad
+                                              origin:SentryTraceOrigin.autoUIViewController
                                              inBlock:callbackToOrigin];
 
             SentrySpanId *layoutSubViewId =
                 [self.tracker startSpanWithName:@"layoutSubViews"
                                      nameSource:kSentryTransactionNameSourceComponent
-                                      operation:SentrySpanOperationUILoad
-                                         origin:SentryTraceOriginAutoUIViewController];
+                                      operation:SentrySpanOperation.uiLoad
+                                         origin:SentryTraceOrigin.autoUIViewController];
 
             objc_setAssociatedObject(controller,
                 &SENTRY_UI_PERFORMANCE_TRACKER_LAYOUTSUBVIEW_SPAN_ID, layoutSubViewId,
@@ -332,8 +356,9 @@
         // (https://developer.apple.com/documentation/uikit/uiviewcontroller/1621510-viewwillappear),
         // viewWillAppear should be called for before the UIViewController is added to the view
         // hierarchy. There are some edge cases, though, when this doesn't happen, and we saw
-        // customers' transactions also proofing this. Therefore, we must also report the initial
-        // display here, as the customers' transactions had spans for `viewWillLayoutSubviews`.
+        // customers' transactions also proofing this. Therefore, we must also report the
+        // initial display here, as the customers' transactions had spans for
+        // `viewWillLayoutSubviews`.
 
         [self reportInitialDisplay:controller];
     };
@@ -368,8 +393,8 @@
 
             [self.tracker measureSpanWithDescription:@"viewDidLayoutSubviews"
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
-                                              origin:SentryTraceOriginAutoUIViewController
+                                           operation:SentrySpanOperation.uiLoad
+                                              origin:SentryTraceOrigin.autoUIViewController
                                              inBlock:callbackToOrigin];
 
             objc_setAssociatedObject(controller,
@@ -387,9 +412,10 @@
 }
 
 /**
- * When a custom UIViewController is a subclass of another custom UIViewController, the SDK swizzles
- * both functions, which would create one span for each UIViewController leading to duplicate spans
- * in the transaction. To fix this, we only allow one span per lifecycle method at a time.
+ * When a custom UIViewController is a subclass of another custom UIViewController, the SDK
+ * swizzles both functions, which would create one span for each UIViewController leading to
+ * duplicate spans in the transaction. To fix this, we only allow one span per lifecycle method
+ * at a time.
  */
 - (void)limitOverride:(NSString *)description
                target:(UIViewController *)viewController
@@ -430,8 +456,8 @@
     } else {
         [self.tracker measureSpanWithDescription:description
                                       nameSource:kSentryTransactionNameSourceComponent
-                                       operation:SentrySpanOperationUILoad
-                                          origin:SentryTraceOriginAutoUIViewController
+                                       operation:SentrySpanOperation.uiLoad
+                                          origin:SentryTraceOrigin.autoUIViewController
                                     parentSpanId:spanId
                                          inBlock:callbackToOrigin];
     }