+++ Sources/Sentry/include/SentryPrivate.h
@@ -14,6 +14,7 @@
 #import "SentryDisplayLinkWrapper.h"
 #import "SentryLevelHelper.h"
 #import "SentryLogC.h"
+#import "SentryMeta.h"
 #import "SentryRandom.h"
 #import "SentrySdkInfo.h"
 #import "SentrySession.h"