+++ Tests/SentryTests/Integrations/Performance/IO/SentryFileIOTrackingIntegrationObjCTests.m
@@ -70,7 +70,7 @@
 
 - (void)test_dataWithContentsOfFile
 {
-    [self assertTransactionForOperation:SENTRY_FILE_READ_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileRead
                                   block:^{
                                       [self assertData:[NSData
                                                            dataWithContentsOfFile:self->filePath]];
@@ -80,7 +80,7 @@
 - (void)test_dataWithContentsOfFileOptionsError
 {
     [self
-        assertTransactionForOperation:SENTRY_FILE_READ_OPERATION
+        assertTransactionForOperation:SentrySpanOperation.fileRead
                                 block:^{
                                     [self
                                         assertData:[NSData
@@ -93,7 +93,7 @@
 - (void)test_dataWithContentsOfURL
 {
     [self
-        assertTransactionForOperation:SENTRY_FILE_READ_OPERATION
+        assertTransactionForOperation:SentrySpanOperation.fileRead
                                 block:^{
                                     [self assertData:[NSData dataWithContentsOfURL:self->fileUrl]];
                                 }];
@@ -102,7 +102,7 @@
 - (void)test_dataWithContentsOfURLOptionsError
 {
     [self
-        assertTransactionForOperation:SENTRY_FILE_READ_OPERATION
+        assertTransactionForOperation:SentrySpanOperation.fileRead
                                 block:^{
                                     [self assertData:[NSData
                                                          dataWithContentsOfURL:self->fileUrl
@@ -113,7 +113,7 @@
 
 - (void)test_initWithContentsOfURL
 {
-    [self assertTransactionForOperation:SENTRY_FILE_READ_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileRead
                                   block:^{
                                       [self assertData:[[NSData alloc]
                                                            initWithContentsOfURL:self->fileUrl]];
@@ -122,7 +122,7 @@
 
 - (void)test_initWithContentsOfFile
 {
-    [self assertTransactionForOperation:SENTRY_FILE_READ_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileRead
                                   block:^{
                                       [self assertData:[[NSData alloc]
                                                            initWithContentsOfFile:self->filePath]];
@@ -131,7 +131,7 @@
 
 - (void)test_writeToFileAtomically
 {
-    [self assertTransactionForOperation:SENTRY_FILE_WRITE_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileWrite
                                   block:^{
                                       [self->someData writeToFile:self->filePath atomically:true];
                                   }];
@@ -140,7 +140,7 @@
 
 - (void)test_writeToUrlAtomically
 {
-    [self assertTransactionForOperation:SENTRY_FILE_WRITE_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileWrite
                                   block:^{
                                       [self->someData writeToURL:self->fileUrl atomically:true];
                                   }];
@@ -149,7 +149,7 @@
 
 - (void)test_writeToFileOptionsError
 {
-    [self assertTransactionForOperation:SENTRY_FILE_WRITE_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileWrite
                                   block:^{
                                       [self->someData writeToFile:self->filePath
                                                           options:NSDataWritingAtomic
@@ -160,7 +160,7 @@
 
 - (void)test_writeToUrlOptionsError
 {
-    [self assertTransactionForOperation:SENTRY_FILE_WRITE_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileWrite
                                   block:^{
                                       [self->someData writeToURL:self->fileUrl
                                                          options:NSDataWritingAtomic
@@ -171,7 +171,7 @@
 
 - (void)test_NSFileManagerContentAtPath
 {
-    [self assertTransactionForOperation:SENTRY_FILE_READ_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileRead
                                   block:^{
                                       [self assertData:[NSFileManager.defaultManager
                                                            contentsAtPath:self->filePath]];
@@ -180,7 +180,7 @@
 
 - (void)test_NSFileManagerCreateFile
 {
-    [self assertTransactionForOperation:SENTRY_FILE_WRITE_OPERATION
+    [self assertTransactionForOperation:SentrySpanOperation.fileWrite
                                   block:^{
                                       [NSFileManager.defaultManager createFileAtPath:self->filePath
                                                                             contents:self->someData
@@ -219,7 +219,7 @@
 
     NSString *filename = filePath.lastPathComponent;
 
-    if ([operation isEqualToString:SENTRY_FILE_READ_OPERATION]) {
+    if ([operation isEqualToString:SentrySpanOperation.fileRead]) {
         XCTAssertEqualObjects(ioSpan.spanDescription, filename);
     } else {
         NSString *expectedString = [NSString stringWithFormat:@"%@ (%@)", filename,